+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [2.2.3](https://github.com/Okipa/laravel-bootstrap-components/compare/2.2.2...2.2.3)
+
+2020-07-16
+
+* Fixed form components label hiding when setting `null` or `false`.
+* Fixed form components `prepend` and `append` methods in order to let them accept `false` to hide them.
+
 ## [2.2.2](https://github.com/Okipa/laravel-bootstrap-components/compare/2.2.1...2.2.2)
 ## [2.2.1](https://github.com/Okipa/laravel-bootstrap-components/compare/2.2.0...2.2.1)